+++ packages/common/ProForm/src/types/contextTypes.ts
@@ -1,4 +1,8 @@
-import { AdaptedInterface, NativeCustomizationOptions } from ".";
+import {
+	AdaptedInterface,
+	DisplayOptions,
+	NativeCustomizationOptions,
+} from ".";
 import { AnyObject } from "./utilTypes";
 
 export type DomType = new (...args: any) => AnyObject & {
@@ -18,6 +22,7 @@ export interface FormPreset {
 	};
 	adapter?: AdaptedInterface;
 	native?: NativeCustomizationOptions;
+	display?: DisplayOptions;
 }
 
 export type UIName = "ArcoVue" | "NutUI" | "NaiveUI" | (string & {});