+++ packages/common/ProForm/src/types/formCustomizerTypes.ts
@@ -148,3 +148,7 @@ export type NativeCustomizationOptions = {
 		FormItem?: AnyObject;
 	};
 };
+
+export type DisplayOptions = {
+	labelPrefix: string;
+};